FIX stop using expired deprecation for numpy 1.24 (#479)
* MNT stop using np.int and np.float

* FIX use array of objects to allow different lengths for each hemisphere

* WIP debug SyntaxError: not a PNG file

* WIP

* WIP

* WIP

* FIX use integer height parameter for Inkscape

+++ cortex/webgl/serve.py
@@ -38,9 +38,9 @@ def make_base64(imgfile):
 class NPEncode(json.JSONEncoder):
     def default(self, obj):
         if isinstance(obj, np.ndarray):
-            if obj.dtype == np.float:
+            if obj.dtype == float:
                 obj = obj.astype(np.float32)
-            elif obj.dtype == np.int:
+            elif obj.dtype == int:
                 obj = obj.astype(np.int32)
 
             return dict(
